separated draw script
